update the integer id to long id

+++ blog-portal/src/main/java/site/iblogs/portal/model/request/CommentRequest.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class CommentRequest {
-    private Integer parentId;
-    private int contentId;
+    private Long parentId;
+    private Long contentId;
     private String author;
     private String mail;
     private String url;
